Add APN push private key.
Normally this would be provided by Apple, but we aren't planning on
sending any notifications so we just need to provide a valid RSA private
key in order to initialize the APNSender. I could also completely remove
the APNSender but let's see if it's possible to get this running with as
little code changes as possible.